move id types into a idtypes module
