+++ src/model/enums/mod.rs
@@ -1,6 +1,5 @@
 //! All Enums for Rspotify
 pub mod country;
-pub mod idtypes;
 pub mod misc;
 pub mod types;
